+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @design-sync/sync
 
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @design-sync/w3c-dtfm@0.9.1
+  - @design-sync/manager@0.5.2
+
 ## 0.9.1
 
 ### Patch Changes